OHC-2: Implement CSV parsing and data loading

+++ src/main/java/com/epam/swissre/interview/orghierarchy/model/Employee.java
@@ -3,7 +3,7 @@ package com.epam.swissre.interview.orghierarchy.model;
 /**
  * Represents an employee with basic details including their manager.
  */
-public record Employee(int id, String firstName, String lastName, double salary, Integer managerId) {
+public record Employee(int id, String firstName, String lastName, int salary, Integer managerId) {
   public Employee {
     if (id <= 0) {
       throw new IllegalArgumentException("ID must be positive.");
